fix clean to avoid deleting protected names

+++ siliconcompiler/scheduler/scheduler.py
@@ -534,6 +534,8 @@ class Scheduler:
             for delfile in os.listdir(cur_job_dir):
                 if delfile == "job.log" and recheck:
                     continue
+                if delfile.startswith("sc_") and recheck:
+                    continue
                 if os.path.isfile(os.path.join(cur_job_dir, delfile)):
                     os.remove(os.path.join(cur_job_dir, delfile))
                 else:
@@ -548,19 +550,17 @@ class Scheduler:
         node's runtime context and invokes its clean_directory method to perform
         node-specific cleanup.
         """
-        protected_dirs = {os.path.basename(collectiondir(self.__project))}
-
         keep_steps = set([step for step, _ in self.__flow.get_nodes()])
         cur_job_dir = jobdir(self.__project)
         for step in os.listdir(cur_job_dir):
-            if step in protected_dirs:
+            if step.startswith("sc_"):
                 continue
             if not os.path.isdir(os.path.join(cur_job_dir, step)):
                 continue
             if step not in keep_steps:
                 shutil.rmtree(os.path.join(cur_job_dir, step))
         for step in os.listdir(cur_job_dir):
-            if step in protected_dirs:
+            if step.startswith("sc_"):
                 continue
             if not os.path.isdir(os.path.join(cur_job_dir, step)):
                 continue
